rearanged folder & files, fix websockets error, complete feedback for interview
