+++ snapshot-react/package.json
@@ -18,7 +18,6 @@
     "@mui/x-tree-view": "^7.3.0",
     "@react-oauth/google": "^0.12.1",
     "@reduxjs/toolkit": "^2.2.3",
-    "@types/sockjs-client": "^1.5.4",
     "axios": "^1.6.8",
     "d3": "^7.9.0",
     "husky": "^9.0.11",
@@ -42,6 +41,7 @@
     "@types/node": "^20.12.7",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
+    "@types/sockjs-client": "^1.5.4",
     "@types/stompjs": "^2.3.9",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
     "@typescript-eslint/parser": "^7.2.0",
