+++ snapshot-react/src/App.tsx
@@ -6,10 +6,8 @@ import Layout from './components/Layout/Layout';
 import OAuth2RedirectHandler from './pages/Auth/OAuth2RedirectHandler';
 import SignInPage from './pages/Auth/SignInPage';
 import SignUpPage from './pages/Auth/SignUpPage';
-import Feedback from './pages/Feedback/Feedback';
 import Home from './pages/Home/Home';
 import InterviewPage from './pages/Interview/InterviewPage';
-import InterviewModal from './pages/Profile/InterviewModal';
 import MyInterviews from './pages/Profile/MyInterviews';
 import Profile from './pages/Profile/Profile';
 import ProtectedRoute from './utils/ProtectedRoute';
@@ -23,13 +21,10 @@ function App(): JSX.Element {
         <Route path="sign-up" element={<SignUpPage />} />
         <Route path="oauth2/redirect" element={<OAuth2RedirectHandler />} />
         <Route element={<ProtectedRoute />}>
-          <Route path="feedback" element={<Feedback />} />
           <Route path="interview" element={<InterviewPage />} />
           <Route path="interview/:id" element={<InterviewPage />} />
-          <Route path="profile" element={<Profile />}>
-            <Route path="interview-journal" element={<MyInterviews />}>
-              <Route path=":interviewId" element={<InterviewModal />} />
-            </Route>
+          <Route path="profile/:userId" element={<Profile />}>
+            <Route path="interview-journal" element={<MyInterviews />} />
           </Route>
         </Route>
         <Route path="*" element={<p>404</p>} />
