--- snapshot-react/src/models/feedback/IFeedback..ts
@@ -1,8 +1,0 @@
-import Question from './IQuestion';
-
-interface IFeedback {
-  questions: Question[],
-  feedback: string,
-}
-
-export default IFeedback;
