--- snapshot-react/src/models/feedback/IFeedbackItem.ts
@@ -1,7 +1,0 @@
-interface IFeedbackItem {
-  question: string,
-  score: string,
-  skillName: string,
-}
-
-export default IFeedbackItem;
