+++ snapshot-react/src/models/profile/IQuestion.ts
@@ -3,6 +3,7 @@ interface IQuestion {
   question: string;
   grade: string;
   skillName: string;
+  searcherId: number;
 }
 
 export default IQuestion;
