+++ snapshot-react/src/pages/Profile/components/Skills/Skill.module.scss
@@ -1,6 +1,5 @@
-.skill{
-  border-radius:10px;
+.skill {
   padding: 5px;
   background: #636b74;
-
+  border-radius: 10px;
 }
