+++ snapshot-react/src/pages/Profile/components/Skills/Skill.tsx
