--- snapshot-react/src/store/reducers/feedback/common.ts
@@ -1,5 +1,0 @@
-enum ActionType {
-  GET_FEEDBACK = 'feedback/getFeedback',
-}
-
-export default ActionType;
